app/requirements.txt
falcon==3.1.0
uvicorn==0.18.3
asyncio==3.4.3
aiofiles==0.8.0
Pillow==9.1.1
mongoengine==0.24.2
PyJWT==2.6.0